Implemented LikePost and LikeComment Model

+++ src/main/java/com/charlesbabbage/fashionblogapi/serviceImpl/CommentServiceImpl.java
@@ -1,16 +1,18 @@
 package com.charlesbabbage.fashionblogapi.serviceImpl;
 
 import com.charlesbabbage.fashionblogapi.dto.CommentDTO;
-import com.charlesbabbage.fashionblogapi.exception.ResourceNotFoundException;
 import com.charlesbabbage.fashionblogapi.model.Comment;
 import com.charlesbabbage.fashionblogapi.model.Post;
 import com.charlesbabbage.fashionblogapi.model.User;
+import com.charlesbabbage.fashionblogapi.pojos.APIResponse;
 import com.charlesbabbage.fashionblogapi.repository.CommentRepository;
 import com.charlesbabbage.fashionblogapi.repository.PostRepository;
 import com.charlesbabbage.fashionblogapi.repository.UserRepository;
 import com.charlesbabbage.fashionblogapi.service.CommentService;
+import com.charlesbabbage.fashionblogapi.utils.ResponseUtil;
 import com.charlesbabbage.fashionblogapi.utils.UUID;
 import lombok.AllArgsConstructor;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
 
@@ -22,8 +24,10 @@ public class CommentServiceImpl implements CommentService {
     CommentRepository commentRepo;
     PostRepository postRepo;
 
+    ResponseUtil responseUtil;
+
     @Override
-    public Comment createComment(CommentDTO commentDTO) {
+    public ResponseEntity<APIResponse> createComment(CommentDTO commentDTO) {
         String id = UUID.getUniqueId();
         Post post = postRepo.findById(commentDTO.getPost_id()).get();
         User user = userRepo.findById(commentDTO.getUser_id()).get();
@@ -32,25 +36,32 @@ public class CommentServiceImpl implements CommentService {
         comment.setComment(commentDTO.getComment());
         comment.setPost(post);
         comment.setUser(user);
-        return commentRepo.save(comment);
+        return responseUtil.Okay(commentRepo.save(comment));
     }
 
     @Override
-    public Comment getComment(String id) {
-        return commentRepo.findById(id).orElseThrow(()-> new ResourceNotFoundException("Comment","id",id));
-    }
+    public ResponseEntity<APIResponse> getComment(String id) {
+        if (commentRepo.findById(id).isEmpty()){
+            return responseUtil.NotFound("Comment bot found");
+        }
+        return responseUtil.Okay(commentRepo.findById(id).get());  }
 
     @Override
-    public Comment editComment(String message, String id) {
-        Comment comment = getComment(id);
+    public ResponseEntity<APIResponse> editComment(String message, String id) {
+        if (commentRepo.findById(id).isEmpty()){
+            return responseUtil.NotFound("Comment not found");
+        }
+        Comment comment = commentRepo.findById(id).get();
         comment.setComment(message);
-        return commentRepo.save(comment);
+        return responseUtil.Okay(commentRepo.save(comment));
     }
 
     @Override
-    public String deleteComment(String id) {
-        getComment(id);
+    public ResponseEntity<APIResponse> deleteComment(String id) {
+        if (commentRepo.findById(id).isEmpty()){
+            return responseUtil.NotFound("Comment not found");
+        }
         commentRepo.deleteById(id);
-        return "deleted";
+        return responseUtil.Deleted();
     }
 }
